Move dev command to folder
